Moved Piwik\TaskScheduler to Piwik\Scheduler\Scheduler
